git-svn-id: https://mseide-msegui.svn.sourceforge.net/svnroot/mseide-msegui/trunk@3555 56ecf5fe-6917-0410-8ac6-ca8e178bccbf

+++ lib/common/kernel/i386-linux/msesysintf.pas
@@ -961,7 +961,8 @@ end;
 
 function filetimetodatetime(sec: time_t; nsec: longword): tdatetime;
 begin
- result:= sec / (24.0*60.0*60.0) + nsec / (24.0*60.0*60.0*1e9) - datetimeoffset;
+ result:= sec / (double(24.0)*60.0*60.0) + 
+          nsec / (double(24.0)*60.0*60.0*1e9) - datetimeoffset;
 end;
 
 function sys_getcurrentdir: msestring;
